qt changes

+++ src/MEGASync/gui/ButtonIconManager.cpp
@@ -234,7 +234,7 @@ void ButtonIconManager::changeButtonTextColor(QAbstractButton* button, double al
             || !button->property(NOT_CHANGE_TEXT_COLOR).isValid()))
     {
         QColor textColor(button->palette().color(QPalette::ColorRole::ButtonText));
-        textColor.setAlphaF(static_cast<float>(alphaValue));
+        textColor.setAlphaF(alphaValue);
         auto paletteButton = button->palette();
         paletteButton.setColor(QPalette::ButtonText, textColor);
         button->setPalette(paletteButton);
